Modularize Queue
