+++ message-processor/media.py
@@ -1,0 +1,4 @@
+from lib.queue import q
+
+def media_file_downloaded(json_string):
+    q.enqueue('audio.new_file_present', json_string)
